requirements.txt
Flask==2.3.3
google-genai==1.17.0
python-dotenv==1.0.0